dsa
das

+++ commands/die.js
@@ -3,13 +3,13 @@ const Discord = require('discord.js');
 const { MessageEmbed } = require("discord.js");
 
 module.exports = {
-	name: 'die',
-	description: 'die',
+	name: 'perish',
+	description: 'perish',
 	execute(message, args) {
 
-		//if (message.member.hasPermission('MANAGE_CHANNELS')) {
+		if (message.member.hasPermission('MANAGE_CHANNELS')) {
 			message.channel.delete()
-	//	}
+		}
 		
     },
 };
